fix(kube): guard the pod count's metadata-only request, not just its answer
The test named for the property it was supposed to protect didn't check
it: swapping list_metadata for list left every assertion green, because
a full PodList deserializes fine into the metadata-only shape and
produces the same counts from the same fixture — it just ships every
pod's containers, images and spec/status across the wire to get there.

The fake server now records each request's Accept header and the test
asserts it names PartialObjectMetadataList; reverting to list fails it.

Also moves the capability out of metrics.rs into its own module — that
file's header specifically describes metrics.k8s.io ("returns an error
if metrics-server is not installed"), which pod count neither reaches
nor depends on — and fixes an import that had drifted out of
alphabetical order along the way.
