+++ crates/kube/src/lib.rs
@@ -123,6 +123,7 @@ pub mod networkpolicies;
 pub mod nodes;
 pub mod oidc_detect;
 pub mod persistentvolumes;
+pub mod pod_count;
 pub mod pvcs;
 pub mod resourcequotas;
 pub mod rolebindings;
