+++ crates/registry/src/lib.rs
@@ -337,7 +337,7 @@ pub fn build_registry_with_paths_and_settings(
         cache.clone(),
     ));
     reg.register(srelens_kube::metrics::pod_metrics_capability(cache.clone()));
-    reg.register(srelens_kube::metrics::pod_count_capability(cache.clone()));
+    reg.register(srelens_kube::pod_count::pod_count_capability(cache.clone()));
     reg.register(srelens_kube::nodes::list_nodes_capability(cache.clone()));
     reg.register(srelens_kube::manifest::get_manifest_capability(
         cache.clone(),
